requirements.txt
flask[async]>=2.2.0
werkzeug>=2.2.0
flask-cors==3.0.10
gunicorn==20.1.0

# Google Cloud Services
google-cloud-storage==2.5.0
google-cloud-bigquery>=3.4.0
google-cloud-speech==2.11.0
google-cloud-texttospeech>=2.12.0
google-cloud-aiplatform>=1.38.0
google-generativeai>=0.3.0
google-genai
google-cloud-tasks==2.13.1

# Banco de dados e manipulação de dados
pydantic>=2.0.0
unidecode>=1.3.0

# IA e Machine Learning
vertexai>=0.0.1
diffusers>=0.21.0
torch>=2.0.0
openai>=1.0.0

# Processamento de imagem e vídeo
opencv-python-headless>=4.8.0
ffmpeg-python==0.2.0
Pillow>=10.0.0  # Para manipulação de imagens com PIL

# Reconhecimento de Fala e TTS
SpeechRecognition>=3.10.0
edge-tts==6.1.3
gTTS>=2.3.0

# Requests e comunicação com APIs
requests>=2.31.0
asyncio
aiohttp>=3.8.0
translate>=3.6.1  # Substituindo googletrans por uma biblioteca mais estável

# Utilitários
python-dotenv>=1.0.0
numpy<2  # Evitar conflitos com pyarrow e Google Cloud SDK
pyarrow<11.0.0,>=3.0.0  # Versão compatível com google-cloud-bigquery
protobuf>=3.19.0,<4.0.0  # Versão compatível com google-cloud-texttospeech
aiofiles>=23.0.0  # Para manipulação assíncrona de arquivos
jinja2>=3.0.0  # Para templates HTML
PyPDF2>=3.0.0  # Para manipulação de PDFs

# Documentação
pdoc3==0.10.0
sphinx==7.1.2
sphinx-rtd-theme==1.3.0

# Dependências do Sistema (precisam ser instaladas separadamente)
# sudo apt-get install ffmpeg

# Áudio e Mídia
pydub>=0.25.1
weasyprint>=60.0  # Para geração de PDFs

# Google APIs e Autenticação
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

pyOpenSSL>=23.0.0
certifi>=2024.2.0

psutil==5.9.8
google-cloud-logging